chore: extract snyk-monitor deployment in tests
Similarly to our platforms module, create a module that handles the deployment of the snyk-monitor.
Today there are two ways to deploy the product: with YAMLs and with a Helm chart.
The goal of this deployment module is to add support for Helm deployments in the future, and for Operators.
This is the first step of organising the code, preserving the original flow, without extra refactoring.
