+++ test/setup/deployers/types.ts
@@ -1,0 +1,5 @@
+export enum DeploymentType {
+  YAML,
+  Helm,
+  Operator,
+}
